requirements.txt
numpy
requests
pandas
websockets
pyyaml
